backend/python/nexdefend-ai/requirements.txt
Flask
numpy
scikit-learn
python-nmap
requests
